% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb}
\alias{cwb}
\title{Calculate p-values with cluster wild bootstrapping for meta-regressions}
\usage{
cwb(
  dat,
  smd,
  var,
  cluster,
  full_model,
  null_model,
  indices,
  R = 999,
  adjust = FALSE
)
}
\arguments{
\item{dat}{data frame or tibble containing effect sizes, variance of effect sizes, study identification number, and moderating variables.}

\item{smd}{name of the column containing the standardized mean differences.}

\item{var}{name of the column containing the variances of the standardized mean differences.}

\item{cluster}{name of the column containing study identifiers.}

\item{full_model}{a model fit using \code{robu()} that includes all moderators of interest.}

\item{null_model}{a model fit using \code{robu()} that includes only the moderators from the full model that are not to be tested.}

\item{indices}{indices of the variables to be tested in single coefficient test or multiple contrast hypothesis test.}

\item{R}{number of bootstrap replications.}

\item{adjust}{logical indicating whether or not to multiply residuals by CR2 adjustment matrices when bootstrapping.}
}
\value{
A tibble containing the name of the test and a p-value.
}
\description{
Calculates p-values for single coefficient and multiple contrast hypothesis tests using cluster wild bootstrapping.
}
\examples{
library(clubSandwich)
library(robumeta)

full <- robu(d ~ study_type,
             studynum = study,
             var.eff.size = V,
             small = FALSE,
             data = SATcoaching)

null <- robu(d ~ 1,
             studynum = study,
             var.eff.size = V,
             small = FALSE,
             data = SATcoaching)

cwb(dat = SATcoaching,
    smd = d,
    var = V,
    cluster = study,
    full_model = full,
    null_model = null,
    indices = 2:3,
    R = 99)

}
