% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boot_distribution.R
\name{plot_boot_distribution}
\alias{plot_boot_distribution}
\title{Plot distribution of bootstrap replicates}
\usage{
plot_boot_distribution(boot_dat, ...)
}
\arguments{
\item{boot_dat}{results from the cwb() function.}
}
\value{
A density plot showing the distribution of bootstrap replicates.
}
\description{
Plots distributions of bootstrap replicates, i.e, meta-regression coefficients obtained from different bootstrap samples.
}
\examples{
library(clubSandwich)
library(robumeta)

full <- robu(d ~ study_type,
             studynum = study,
             var.eff.size = V,
             small = FALSE,
             data = SATcoaching)

cwb_dat <- cwb(full_model = full,
               indices = 2:3,
               R = 99)

plot_boot_distribution(boot_dat = cwb_dat)

}
