% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wildmeta.R
\name{plot_cwb.Wald_test_wildmeta}
\alias{plot_cwb.Wald_test_wildmeta}
\title{Plot distribution of bootstrap test statistics}
\usage{
\method{plot_cwb}{Wald_test_wildmeta}(results, ...)
}
\arguments{
\item{results}{Results from Wald_test_cwb function}

\item{...}{Any other arguments to be passed to \code{ggplot2::geom_density())}}
}
\value{
A ggplot2 density plot.
}
\description{
Creates a density plot showing the distribution of bootstrap test statistics.
}
\examples{
library(clubSandwich)
library(robumeta)

model <- robu(d ~ 0 + study_type + hrs + test,
             studynum = study,
              var.eff.size = V,
              small = FALSE,
              data = SATcoaching)

res <- Wald_test_cwb(full_model = full_model,
                     constraint_matrix = C_mat,
                     R = 12)

plot_cwb(res, fill = "darkred", alpha = 0.5)

}
