% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{constrain_equal}
\alias{constrain_zero}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{clubSandwich}{\code{\link[clubSandwich:constraint_matrices]{constrain_equal}}, \code{\link[clubSandwich:constraint_matrices]{constrain_zero}}}
}}

