% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb}
\alias{cwb}
\title{Calculate p-values with cluster wild bootstrapping for meta-regressions}
\usage{
cwb(
  dat,
  smd,
  var,
  cluster,
  covs_full_form,
  test_vars,
  indices,
  R = 999,
  adjust = FALSE
)
}
\arguments{
\item{dat}{data frame or tibble containing effect sizes, variance of effect sizes, study identification number, and moderating variables.}

\item{smd}{name of the column containing the standardized mean differences.}

\item{var}{name of the column containing the variances of the standardized mean differences.}

\item{cluster}{name of the column containing study identifiers.}

\item{covs_full_form}{a string containing the right hand side of the meta-regression model (e.g., "X1 + X2 + X3 + X4 + X5").}

\item{test_vars}{a string or character vector containing the variables to be tested.}

\item{indices}{indices of the variables to be tested in single coefficient test or multiple contrast hypothesis test.}

\item{R}{number of bootstrap replications.}

\item{adjust}{logical indicating whether or not to multiply residuals by CR2 adjustment matrices when bootstrapping.}
}
\value{
A tibble containing the name of the test and a p-value.
}
\description{
Calculates p-values for single coefficient and multiple contrast hypothesis tests using cluster wild bootstrapping.
}
\examples{
library(clubSandwich)
cwb(dat = SATcoaching,
    smd = d,
    var = V,
    cluster = study,
    covs_full_form = "study_type",
    test_vars = "study_type",
    indices = 2:3,
    R = 99)

}
