% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb}
\alias{cwb}
\title{Calculate p-values with cluster wild bootstrapping for meta-regressions}
\usage{
cwb(data, smd, var, cluster, full_model, null_model, indices, R = 999)
}
\arguments{
\item{data}{data frame or tibble containing effect sizes, variance of effect sizes, study identification number, and moderating variables.}

\item{smd}{name of the column containing the standardized mean differences.}

\item{var}{name of the column containing the variances of the standardized mean differences.}

\item{cluster}{name of the column containing study identifiers.}

\item{full_model}{full model fit using robumeta or metafor.}

\item{null_model}{null model fit using robumeta or metafor. The null model includes all covariates from the full model that are not to be tested.}

\item{indices}{indices of the variables to be tested in single coefficient test or multiple contrast hypothesis test.}

\item{R}{number of bootstrap replications.}
}
\value{
A tibble containing the p-value.
}
\description{
Calculates p-values for single coefficient and multiple contrast hypothesis tests using cluster wild bootstrapping.
}
\examples{



}
