% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb}
\alias{cwb}
\title{Calculate p-values with cluster wild bootstrapping for meta-regressions}
\usage{
cwb(full_model, indices, R = 999, adjust = FALSE)
}
\arguments{
\item{full_model}{a model fit using \code{robu()} or \code{rma.mv()} that includes all moderators of interest.}

\item{indices}{indices of the variables to be tested in single coefficient test or multiple contrast hypothesis test.}

\item{R}{number of bootstrap replications.}

\item{adjust}{logical indicating whether or not to multiply residuals by CR2 adjustment matrices when bootstrapping.}
}
\value{
A tibble containing the name of the test, p-value, and the name of the working model. CE denotes correlated effects and HE denotes hierarchical effects.
}
\description{
Calculates p-values for single coefficient and multiple contrast hypothesis tests using cluster wild bootstrapping.
}
\examples{
library(clubSandwich)
library(robumeta)

full <- robu(d ~ study_type,
             studynum = study,
             var.eff.size = V,
             small = FALSE,
             data = SATcoaching)

cwb(full_model = full,
    indices = 2:3,
    R = 99)

}
